requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
python-dotenv
Flask-WTF
pytest
